Refactors most of azure_core to typespec
Little was modified when moving types and functions. Of what little was modified, most types were either separated from unrelated topics - like streams from HTTP - or better consolidated into fewer submodules to avoid complicating the module hierarchy.

`TelemetryOptions` was temporarily removed. Because `ClientOptions` was moved to `typespec_client_core` but telemetry considered Azure-specific by most other Azure SDK languages, we need to figure out how to pass this in an unbranded way. Since it basically just sets the `User-Agent` HTTP header, perhaps we can make it more generic and optional.
