--- sdk/core/azure_core/src/date/mod.rs
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-//! Azure date and time parsing and formatting
-
-pub use typespec_client_core::date::*;
