--- sdk/core/azure_core/src/error.rs
@@ -1,5 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-pub use typespec::error::*;
-pub use typespec_client_core::error::{http_response_from_body, HttpError};
