--- sdk/core/azure_core/src/json.rs
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-pub use typespec_client_core::json::*;
