+++ sdk/core/azure_core/src/macros.rs
@@ -1,60 +1,3 @@
-/// Creates setter methods
-///
-/// The methods created are of the form `$name` that takes an argument of type `$typ`
-/// and sets the field `$name` to result of calling `$transform` with the value of the argument.
-///
-/// In other words. The following macro call:
-/// ```
-/// # #[macro_use] extern crate azure_core;
-/// struct MyStruct<'a> { foo: Option<&'a str> };
-/// impl <'a> MyStruct<'a> {
-///     setters! { foo: &'a str => Some(foo), }
-/// }
-/// ```
-/// Roughly expands to:
-/// ```
-/// struct MyStruct<'a> { foo: Option<&'a str> };
-/// impl <'a> MyStruct<'a> {
-///     fn foo(self, foo: &'a str) -> Self {
-///         Self {
-///             foo: Some(foo),
-///             ..self
-///         }
-///     }
-/// }
-/// ```
-#[macro_export]
-macro_rules! setters {
-    (@single $(#[$meta:meta])* $name:ident : $typ:ty => $transform:expr) => {
-        #[allow(clippy::redundant_field_names)]
-        #[allow(clippy::needless_update)]
-        #[allow(missing_docs)]
-        #[must_use]
-        $(#[$meta])*
-        pub fn $name<P: ::std::convert::Into<$typ>>(self, $name: P) -> Self {
-            let $name: $typ = $name.into();
-            Self  {
-                $name: $transform,
-                ..self
-            }
-        }
-    };
-    // Terminal condition
-    (@recurse) => {};
-    // Recurse without transform
-    (@recurse $(#[$meta:meta])* $name:ident : $typ:ty, $($tokens:tt)*) => {
-        $crate::setters! { @recurse $(#[$meta])* $name: $typ => $name, $($tokens)* }
-    };
-    // Recurse with transform
-    (@recurse $(#[$meta:meta])* $name:ident : $typ:ty => $transform:expr, $($tokens:tt)*) => {
-        $crate::setters! { @single $(#[$meta])* $name : $typ => $transform }
-        $crate::setters! { @recurse $($tokens)* }
-    };
-    ($($tokens:tt)*) => {
-        $crate::setters! { @recurse $($tokens)* }
-    }
-}
-
 /// Declare a `Future` with the given name
 ///
 /// `Future::Output` will be set to `azure_core::Result<$NAMEResponse>`.
@@ -275,25 +218,6 @@ mod test {
     create_enum!(Colors, (Black, "Black"), (White, "White"), (Red, "Red"));
     create_enum!(ColorsMonochrome, (Black, "Black"), (White, "White"));
 
-    struct Options {
-        a: Option<String>,
-        b: u32,
-    }
-
-    #[allow(dead_code)]
-    impl Options {
-        setters! {
-            a: String => Some(a),
-            b: u32 => b,
-        }
-    }
-
-    impl Default for Options {
-        fn default() -> Self {
-            Options { a: None, b: 1 }
-        }
-    }
-
     #[test]
     fn test_color_parse_1() {
         let color = "Black".parse::<Colors>().unwrap();
@@ -309,13 +233,5 @@ mod test {
     #[test]
     fn test_color_parse_err_1() {
         "Red".parse::<ColorsMonochrome>().unwrap_err();
-    }
-
-    #[test]
-    fn test_setters() {
-        let options = Options::default().a("test".to_owned());
-
-        assert_eq!(Some("test".to_owned()), options.a);
-        assert_eq!(1, options.b);
     }
 }
