+++ sdk/core/azure_core/src/policies/mod.rs
@@ -1,34 +1,7 @@
-mod custom_headers_policy;
-mod retry_policies;
-mod telemetry_policy;
-mod transport;
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
 
-pub use custom_headers_policy::{CustomHeaders, CustomHeadersPolicy};
-pub use retry_policies::*;
-pub use telemetry_policy::*;
-pub use transport::*;
+mod telemetry;
 
-use crate::{Context, RawResponse, Request};
-use async_trait::async_trait;
-use std::sync::Arc;
-
-/// A specialized `Result` type for policies.
-pub type PolicyResult = crate::error::Result<RawResponse>;
-
-/// A pipeline policy.
-///
-/// Policies are expected to modify the request and then call the subsequent policy.
-/// Policies can then inspect the response, potentially signaling failure.
-/// The only runtime enforced check is that the last policy must be a Transport policy. It's up to
-/// the implementer to call the subsequent policy.
-/// The `C` generic represents the *contents* of the AuthorizationPolicy specific of this pipeline.
-#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-pub trait Policy: Send + Sync + std::fmt::Debug {
-    async fn send(
-        &self,
-        ctx: &Context,
-        request: &mut Request,
-        next: &[Arc<dyn Policy>],
-    ) -> PolicyResult;
-}
+pub use telemetry::*;
+pub use typespec_client_core::http::policies::*;
