--- sdk/core/azure_core/src/policies/retry_policies/mod.rs
@@ -1,10 +1,0 @@
-mod exponential_retry;
-mod fixed_retry;
-mod no_retry;
-mod retry_policy;
-
-pub use exponential_retry::*;
-pub use fixed_retry::*;
-pub use no_retry::*;
-pub(crate) use retry_policy::get_retry_after;
-pub use retry_policy::RetryPolicy;
