--- sdk/core/azure_core/src/prelude.rs
@@ -1,5 +1,0 @@
-//! The Azure Core prelude.
-
-pub use crate::request_options::*;
-pub use crate::util::*;
-pub use crate::{AppendToUrlQuery, Context, Continuable, Header};
