+++ sdk/core/azure_core/src/tokio/fs.rs
@@ -1,7 +1,3 @@
-use crate::{
-    request::Body,
-    seekable_stream::{SeekableStream, DEFAULT_BUFFER_SIZE},
-};
 use futures::{task::Poll, Future};
 use std::{cmp::min, io::SeekFrom, pin::Pin, sync::Arc, task::Context};
 use tokio::{
@@ -10,6 +6,11 @@ use tokio::{
     sync::Mutex,
 };
 use tracing::debug;
+use typespec_client_core::{
+    http::Body,
+    setters,
+    stream::{SeekableStream, DEFAULT_BUFFER_SIZE},
+};
 
 #[derive(Debug)]
 pub struct FileStreamBuilder {
