+++ sdk/typespec/typespec_client_core/Cargo.toml
@@ -11,18 +11,40 @@ categories = ["compilers", "development-tools"]
 keywords = ["typespec"]
 
 [dependencies]
+async-trait = { workspace = true }
+base64 = { workspace = true }
 bytes = { workspace = true }
+dyn-clone = { workspace = true }
 futures = { workspace = true }
 http-types = { workspace = true, optional = true }
 pin-project = { workspace = true }
 quick-xml = { workspace = true, optional = true }
+rand = { workspace = true }
+reqwest = { workspace = true, optional = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 time = { workspace = true }
+tokio = { workspace = true, optional = true }
+tracing = { workspace = true }
 typespec = { workspace = true, default-features = false }
+url = { workspace = true }
+
+[dev-dependencies]
+once_cell = { workspace = true }
 
 [features]
-default = ["http", "json"]
+default = [
+  "http",
+  "json",
+  "reqwest",
+  "reqwest_gzip",
+  "reqwest_rustls",
+  "tokio_sleep",
+]
 http = ["dep:http-types", "typespec/http"]
 json = ["typespec/json"]
+reqwest = ["dep:reqwest", "reqwest/default-tls"]
+reqwest_gzip = ["reqwest/gzip"]
+reqwest_rustls = ["reqwest/rustls-tls"]
+tokio_sleep = ["dep:tokio"]
 xml = ["dep:quick-xml"]
