+++ sdk/typespec/typespec_client_core/src/http/clients/mod.rs
@@ -1,23 +1,30 @@
-#[cfg(not(any(feature = "enable_reqwest", feature = "enable_reqwest_rustls")))]
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+//! Built-in HTTP clients.
+
+#[cfg(not(any(feature = "reqwest", feature = "reqwest_rustls")))]
 mod noop;
-#[cfg(any(feature = "enable_reqwest", feature = "enable_reqwest_rustls"))]
+#[cfg(any(feature = "reqwest", feature = "reqwest_rustls"))]
 mod reqwest;
 
-#[cfg(not(any(feature = "enable_reqwest", feature = "enable_reqwest_rustls")))]
+#[cfg(not(any(feature = "reqwest", feature = "reqwest_rustls")))]
 use self::noop::new_noop_client;
-#[cfg(any(feature = "enable_reqwest", feature = "enable_reqwest_rustls"))]
+#[cfg(any(feature = "reqwest", feature = "reqwest_rustls"))]
 use self::reqwest::new_reqwest_client;
 
+use crate::http::{RawResponse, Request};
 use async_trait::async_trait;
 use std::sync::Arc;
+use typespec::error::Result;
 
 /// Create a new [`HttpClient`].
 pub fn new_http_client() -> Arc<dyn HttpClient> {
-    #[cfg(any(feature = "enable_reqwest", feature = "enable_reqwest_rustls"))]
+    #[cfg(any(feature = "reqwest", feature = "reqwest_rustls"))]
     {
         new_reqwest_client()
     }
-    #[cfg(not(any(feature = "enable_reqwest", feature = "enable_reqwest_rustls")))]
+    #[cfg(not(any(feature = "reqwest", feature = "reqwest_rustls")))]
     {
         new_noop_client()
     }
@@ -31,5 +38,5 @@ pub trait HttpClient: Send + Sync + std::fmt::Debug {
     ///
     /// It does not consume the request. Implementors are expected to clone the necessary parts
     /// of the request and pass them to the underlying transport.
-    async fn execute_request(&self, request: &crate::Request) -> crate::Result<crate::RawResponse>;
+    async fn execute_request(&self, request: &Request) -> Result<RawResponse>;
 }
