+++ sdk/typespec/typespec_client_core/src/http/policies/retry/fixed.rs
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 use std::time::Duration;
 
 /// Retry policy with a fixed back-off.
