+++ sdk/typespec/typespec_client_core/src/sleep/mod.rs
@@ -1,20 +1,25 @@
-#[cfg(not(feature = "tokio-sleep"))]
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+//! Sleep functions.
+
+#[cfg(not(feature = "tokio_sleep"))]
 mod thread;
 
-#[cfg(not(feature = "tokio-sleep"))]
+#[cfg(not(feature = "tokio_sleep"))]
 pub use self::thread::{sleep, Sleep};
 
-#[cfg(feature = "tokio-sleep")]
+#[cfg(feature = "tokio_sleep")]
 pub use tokio::time::{sleep, Sleep};
 
 // Unit tests
 #[cfg(test)]
 mod tests {
 
-    /// Basic test that launches 10k futures and waits for them to complete
-    /// Has a high chance of failing if there is a race condition in sleep method
-    /// Runs quickly otherwise
-    #[cfg(not(feature = "tokio-sleep"))]
+    // Basic test that launches 10k futures and waits for them to complete:
+    // it has a high chance of failing if there is a race condition in the sleep method;
+    // otherwise, it runs quickly.
+    #[cfg(not(feature = "tokio+sleep"))]
     #[tokio::test]
     async fn test_timeout() {
         use super::*;
