+++ sdk/typespec/typespec_client_core/src/stream/mod.rs
@@ -1,17 +1,23 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+mod bytes_stream;
+
 use bytes::Bytes;
+pub use bytes_stream::*;
 use dyn_clone::DynClone;
 use futures::{io::AsyncRead, stream::Stream, task::Poll};
 use std::{pin::Pin, task::Context};
+use typespec::error::{Error, ErrorKind, Result};
 
-/// Amount of the stream to buffer in memory during streaming uploads
-pub(crate) const DEFAULT_BUFFER_SIZE: usize = 1024 * 64;
+/// Amount of the stream to buffer in memory during streaming uploads.
+pub const DEFAULT_BUFFER_SIZE: usize = 1024 * 64;
 
-/// Enable a type implementing `AsyncRead` to be consumed as if it were
-/// a `Stream` of `Bytes`.
+/// Enable a type implementing `AsyncRead` to be consumed as if it were a `Stream` of `Bytes`.
 #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
 pub trait SeekableStream: AsyncRead + Unpin + std::fmt::Debug + Send + Sync + DynClone {
-    async fn reset(&mut self) -> crate::error::Result<()>;
+    async fn reset(&mut self) -> Result<()>;
     fn len(&self) -> usize;
 
     fn is_empty(&self) -> bool {
@@ -26,7 +32,7 @@ pub trait SeekableStream: AsyncRead + Unpin + std::fmt::Debug + Send + Sync + Dy
 dyn_clone::clone_trait_object!(SeekableStream);
 
 impl Stream for dyn SeekableStream {
-    type Item = crate::error::Result<Bytes>;
+    type Item = Result<Bytes>;
 
     fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         let mut buffer = vec![0_u8; self.buffer_size()];
@@ -38,8 +44,8 @@ impl Stream for dyn SeekableStream {
                 let bytes = bytes.slice(0..bytes_read);
                 Poll::Ready(Some(Ok(bytes)))
             }
-            Poll::Ready(Err(err)) => Poll::Ready(Some(Err(crate::error::Error::full(
-                crate::error::ErrorKind::Io,
+            Poll::Ready(Err(err)) => Poll::Ready(Some(Err(Error::full(
+                ErrorKind::Io,
                 err,
                 "an error was encountered when trying to read from a stream",
             )))),
